+++ app/domain/booking_states/definitive_request.rb
@@ -45,7 +45,6 @@ module BookingStates
       OperatorResponsibility.assign(booking, :home_handover, :home_return)
       MailTemplate.use(:manage_definitive_request_notification, booking, to: :administration, &:autodeliver!)
       mail = MailTemplate.use(:definitive_request_notification, booking, to: :tenant)
-      mail&.attach :accepted_documents if booking.state_transitions.last(2).map(&:to_state) == [OpenRequest.to_s, to_s]
       mail&.autodeliver!
     end