+++ app/services/import/hash/tarif_importer.rb
@@ -8,8 +8,8 @@ module Import
       use_attributes(*%w[associated_types label_i18n ordinal prefill_usage_method price_per_unit tarif_group
                          pin type unit_i18n valid_from valid_until])
 
-      def initialize(organisation, **options)
-        super(**options)
+      def initialize(organisation, **)
+        super(**)
         @organisation = organisation
       end