+++ db/migrate/_cleanup_old_data.rb
@@ -1,0 +1,12 @@
+class CleanupOldData < ActiveRecord::Migration[7.1]
+  def up
+    remove_column :designated_documents, :send_with_contract, :boolean, default: false
+    remove_column :designated_documents, :send_with_last_infos, :boolean, default: false
+    remove_column :designated_documents, :send_with_accepted, :boolean, default: false
+    remove_column :bookings, :booking_questions, :jsonb
+    remove_column :occupiables, :janitor, :text
+    remove_table :bookable_extras
+    remove_table :booked_extras
+
+  end
+end